Created time manager to support search timed interrupt, created a folder structure in sources, fixed a bug which was caused by wrong bitboard size, partial UCI_go handler rewrite
Time managers are now introduced, they will be responsible for search
interruption due to exceeding a time bound. Initial introduction of
BoardInterface has also been made. Bitboard had a size of not 64bit,
when compiling on 64bit Ubuntu using g++ (4.8.3), due to a bad
definition of bitboard which was based on unsigned long long int.
Changing it to uint64_t solved the problem.

Signed-off-by: Chapeiro <periklis93@yahoo.gr>